adding the autoscrolling to the chatBox
