+++ app/src/main/java/com/example/clicker/presentation/stream/views/streamManager/EditStreamInfo.kt
@@ -60,6 +60,7 @@ import androidx.compose.ui.text.input.TextFieldValue
 import androidx.compose.ui.unit.dp
 
 import com.example.clicker.R
+import com.example.clicker.network.models.websockets.TwitchUserData
 import com.example.clicker.presentation.modView.ModViewViewModel
 import com.example.clicker.presentation.sharedViews.SharedComponents
 import com.example.clicker.presentation.stream.FilterType
@@ -97,7 +98,8 @@ fun ManageStreamInformation(
     updateAutoModSettingsStatus:Response<Boolean>?,
     updateAutoModSettingsStatusToNull:()->Unit,
     updateChannelInfo:()->Unit,
-    modViewViewModel: ModViewViewModel
+    modViewViewModel: ModViewViewModel,
+    chatMessages: List<TwitchUserData>
 ){
     if(showAutoModSettings){
 
@@ -125,7 +127,8 @@ fun ManageStreamInformation(
     }else{
         ModView.ModViewScaffold(
             closeStreamInfo={closeStreamInfo()},
-            modViewViewModel =modViewViewModel
+            modViewViewModel =modViewViewModel,
+            chatMessages =chatMessages
         )
     }
